refactor: Use css module in posts page
