+++ src/components/posts/NoResults/NoResults.module.css
@@ -1,0 +1,58 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 18px;
+  margin-top: 60px;
+  margin-bottom: 80px;
+  user-select: none;
+}
+
+.iconContainer {
+  position: relative;
+  display: block;
+  padding: 16px;
+}
+
+.iconContainer::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: block;
+  width: 100%;
+  height: 100%;
+  background-color: var(--theme-posts-noresults-back);
+  border-radius: 10px;
+  transform: rotate(70deg);
+}
+
+.iconContainer::after {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  display: block;
+  width: 100%;
+  height: 100%;
+  border: 3px solid var(--theme-posts-noresults-back);
+  border-radius: 10px;
+  transform: rotate(41deg);
+}
+
+.icon {
+  color: var(--theme-body-text);
+  font-size: 3.4rem;
+}
+
+.aboveIcon {
+  position: absolute;
+  right: -6px;
+  bottom: -4px;
+  font-size: 1.6rem;
+}
+
+.text {
+  font-size: 0.9rem;
+}
