+++ src/components/posts/SearchBox/SearchBox.tsx
@@ -1,6 +1,6 @@
 import { DetailedHTMLProps, InputHTMLAttributes } from 'react'
 import { FiSearch } from 'react-icons/fi'
-import * as S from './SearchBox.styles'
+import S from './SearchBox.module.css'
 
 type Props = DetailedHTMLProps<
   InputHTMLAttributes<HTMLInputElement>,
@@ -13,16 +13,16 @@ export const SearchBox = ({ label, ...otherProps }: Props) => {
   return (
     <>
       <label className="sr-only">{label}</label>
-      <S.SearchBox>
-        <S.SearchBoxIcon as={FiSearch} aria-hidden="true" />
-        <S.SearchBoxInput
+      <div className={S.searchBox}>
+        <FiSearch className={S.searchBoxIcon} aria-hidden="true" />
+        <input className={S.searchBoxInput}
           type="search"
           autoComplete="off"
           name="search"
           placeholder={label}
           {...otherProps}
         />
-      </S.SearchBox>
+      </div>
     </>
   )
 }
